all: refactor to create app in UI
Previously the app was created and passed into the UI, which was a tview
Primitive that could be drawn. Making the UI type act as the application
(instead of just a widget or component) lets us do setup before Run is
called. This also removes the Log option since you could always just
write directly to the pane after creating the application.

Signed-off-by: Sam Whited <sam@samwhited.com>