internal/storage: store bare JID on messages
This is the JID as it would appear in the roster (the bare JID in the
'to' attribute if the message was sent by us, the one in the 'from' attr
otherwise). This can be used to get all messages that should appear in
a conversation, even if the original message was sent from a specific
resourcepart.

Signed-off-by: Sam Whited <sam@samwhited.com>