all: add option to disable TLS
Though this behavior is generally undesirable, it makes it easier to
debug things locally if we have the ability to easily turn off TLS.
It's also possible that this client is used by someone who is securing
their server a different way (ie. by only allowing it to be accessed
over an encrypted VPN).

Though the philosophy of the client is to keep options limited and to
only support features if they cover the 90% use case and not do
"everything and the kitchen sink", I think this option is useful enough
in general that it's worth adding, despite the risks.

Signed-off-by: Sam Whited <sam@samwhited.com>