communique: allow configuring multiple accounts
This allows the config file to store the configuration for multiple
accounts (without letting us actually log in with more than one).
This makes it easier to switch back and forth while debugging without
storing multiple config files, and later may be paired with a flag to
overrid which account to use to make it easy to log in multiple times
without maintaining multiple config files.

Signed-off-by: Sam Whited <sam@samwhited.com>