internal/ui: don't save add roster modal
Instead of saving the add roster modal and reusing it, which causes the
previous state to be saved, re-create it every time that it's used. This
keeps the code all in one places, doesn't clutter up our UI code as
much, and means we dno't need to reset it every time we use it.

Signed-off-by: Sam Whited <sam@samwhited.com>