internal/ui: dont' save quit pane
The quit dialog was previously save and re-used, meaning that if we
closed it and opened it again all state remained the same (eg. the
cancel button might still be selectd). Instead, re-create the quit modal
each time so that we don't keep it around in memory for no reason and so
that the default button is selected each time we start it.

Signed-off-by: Sam Whited <sam@samwhited.com>