communiqué: increase database context timeout
Perhaps due to a slow hard drive, on my machine communiqué always fails
with

	2024/08/15 17:55:08 error opening database: context deadline exceeded

upon first launch (when the database is being initialized). Increase the
corresponding timeout value to avoid the error. As there doesn't seem to
be any obvious reason for why this would even block, set an overly large
limit which nobody should ever hit, but in the unlikely case they do,
the program will gracefully exit and print a useful error message.

Signed-off-by: Karel Balej <balejk@matfyz.cz>