ui: replace emojis with text in the bindings help
The current symbols seem to be broken even in terminals where other
emojis work fine making the help text illegible in places.

While at it, also make it clear that the up/down movement takes
arbitrary count (despite its placement in the help text however, it
seems to only work in the sidebar, not in the chat history, but this is
not addressed here).

Also add the missing bindings for cycling the sidebar tabs.

Signed-off-by: Karel Balej <balejk@matfyz.cz>