db: allow the use of an in-memory database
This was originally meant to be the default for connections using SASL
ANONYMOUS, however it seems like something that could be useful in other
cases as well.

In fact, specifying :memory: without this change already seems to keep
the database in memory, only it also results in an empty ./:memory: file
being created.

Signed-off-by: Karel Balej <balejk@matfyz.cz>