ui: rebind F1 to display the manual page
The original idea was to replace the quick help prompt with this
completely (hence the inaccurate information in the manual page),
however since the manual page may not be available (installed at the
correct location), the quick help is kept for now.

To avoid having to update the bindings in two separate locations, we
could also embed the contents of the manual page in a variable in some
processed format in the future. Also see how for instance catgirl(1),
aerc(1) or lf(1) handle this.

Signed-off-by: Karel Balej <balejk@matfyz.cz>