internal/ui: move status out of conversations list
Previously the status indicator was at the top of the conversations view
in the sidebar. This meant that it was not visible on the roster or
channels views, and that you had to navigate back to the conversations
list or scroll to the top of the list to access it.
It is now above the conversations/roster/channels pager and is always
visible.

Fixes #13

Signed-off-by: Sam Whited <sam@samwhited.com>