internal/ui: refactor keyboard shortcuts
Keyboard shortcuts used to be handled all over the place by individual
widgets. For example, navigation was handled by whatever pane was being
navigated away from, so a "navigate right" might be handled by the
sidebar and "navigate left" might be handled by the logs. This was very
confusing and lead to a lot of inconsistent behavior.

This patch refactors the shortcut handling to move all "global"
shortcuts such as navigation and modal dialogs that should be created
regardless of what is selected (such as the "quick help" dialog) onto
the UI type.

Fixes #27

Signed-off-by: Sam Whited <sam@samwhited.com>