internal/ui: consolidate sidebar shortcuts
These were previously in multiple lists. Instead of splitting them up
based on whether they affect the UI or not, put them all in one place
and pass a pointer to the UI around. This isn't ideal and leads to a
weird pointer loop, but we also aren't going to need to GC the UI or the
sidebar for the life of the application so it's okay until we come up
with a better solution.

Updates #27

Signed-off-by: Sam Whited <sam@samwhited.com>