communiqué: add all new chats to recent chats view
Previously when a new message came in it was only added to the
Conversations view if marking it as unread hadn't worked (ie. if it
didn't already exist). However, this functionality actually indicated
whether it was marked unread in the conversations view *or* the roster
view, so if the contact was in the roster it would never be added to the
Conversations view.

This updates the behavior to always attempt to add the new chat to the
conversations view before marking it as unread.

Fixes #90

Signed-off-by: Sam Whited <sam@samwhited.com>