communiqué: log to stderr until UI is ready
For some reason we setup the logs to point to the UI before the UI was
actually ready, meaning that most log messages would be written to a UI
that wasn't being displayed instead of to stderr. Move swapping over
from early logging to the UI to right before the first time we start the
UI so that all previous messages (if any) will be logged to stderr.

Signed-off-by: Sam Whited <sam@samwhited.com>