communiqué: use CLI subcommands for config file
Previously we generated the config file when the -config flag was
passed. This conflicts with various other flags, so let's clean up the
CLI argument handling to use separate commands such as "config". The
default remains the same (launch the TUI).
This will free us up later to add migrations and other commands in an
easier manner.

Signed-off-by: Sam Whited <sam@samwhited.com>