all: add database migrations script
Previously we were applying the database schema every time we opened the
database and just writing out the schema in such a way that this
wouldn't cause conflicts (ie. using a lot of "IF NOT EXISTS").
This isn't sustainable in the long run, so this patch adds a migrations
system which can be used to keep track of database schema versions and
upgrade (or downgrade) them as the schema evolves.

Fixes #93

Signed-off-by: Sam Whited <sam@samwhited.com>